Solution to Toll Plaza Management System

+++ Toll Plaza Management System@@67877d61dc107486ad5c2750/code.c
@@ -3,7 +3,7 @@
 struct Vehicle{
     char vehicleNumber[100];
     char vehicleType[100];
-    int toll;
+    float toll;
 };
 int main(){
     int n;
